Add tests for events
Currently, tests are failing, because the `adt` library is supposedly
designed to work only with ALL_UPPERCASE case names, and I used
PascalCase names. This wasn't initially clear from the documentation,
and this was reported upstream as an issue:
https://github.com/jspahrsummers/adt/issues/41

The obvious solution is to change case names to ALL_UPPERCASE, which
is probably what I will do soon.