Events: Change name format (PascalCase -> ALL_UPPER) to make tests pass
In event.py and the corresponding part of the test suit,
the format of names of cases of the Event algebraic data type
was changed from PascalCase to ALL_UPPERCASE, so that the `adt`
library can work with them properly. For more information on this
problem, see the previous commit message.