Fix a bug in a unit test.
In test_game.py:TestSnake.test_get_state_copies the `head`
attribute of a snake was mutated and the test checked
that previously returned `SnakeState`s are unaffected by
this. This turned out not to be the case, but the implementation
of `_Snake` was designed to never mutate its `head`. For
the sake of robustness against internal changes, `head`
is now copied when a snapshot of a snake's state is constructed.