Fix misleading variable name
In game_loop.py:run_game_loop the variable name `crashed` was misleading,
because the variable stored True if and only if the snake has *not* crashed.
This was a mistake, but the code worked fine, since another mistake
completely cancelled out the effect of the first one. Now, with
both mistakes fixed, this spot of code is not that misleading anymore.