Fix a type mismatch error in the Field class
Field started to store the objects as a Dict a few commits ago,
but the `get_state` method wasn't changed accordingly. It
simply copied the `_objects` field to include it into the state
object, which no longer was correct when the type of this field
has changed. Now it performs a conversion from a Dict to the
list of tuples.