Make snake generation random
It is implemented as a kind of restricted random walk algorithm,
where the direction of the next tail segment relative to the current
one is chosen at random, but only valid directions are taken into
account. For example, a snake can never turn back or bump into
itself, another snake or the edge of the field, so these directions
are invalid.

It is planned to perform checks whether the snake is locked by itself or
other snakes, which can create trivial and uninteresting battles,
when some snakes simply have no choice but to die. However, the algorithm
for this check may not be trivial (we can run DFS or BFS, but these
algorithms do not take the game dynamics into account), and its
necessity can be questionable (sometimes it might actually be *desired*
to deliberately put bots into these locked-out positions to see how well
they perform in such a setting). As an option, a minimum allowed number
of free cells potentially accessible to snake can be introduced.

To provide a constructive resolution to this issue, this check may
well be made configurable in future releases.