Make turns happen in order
This commit restructures some code as it was deemed necessary during the
development. The role of the scheduler is played with one additional
task running in parallel with the client tasks. It is handled by the
GameRoom class; more precisely, by its method `run_loop`. It iterates
over the players in order provided during GameRoom's construction and
wakes their corresponding client tasks. When a client task signals that
it is waiting for the next turn of the player, it yields to the
scheduler task, which then wakes the client task to take turn next, and
so on. This is repeated in a (currently infinite) loop.

TODO: it is necessary to filter out the dead players, so that they
cannot make any turns, but can send messages to the server and receive
replies. This can be associated and merged with the check for a winner,
so that GameRoom can signal that a game is over.

TODO: currently, messages incoming from the players not in their turns
are not handled and are therefore queued, so that they appear as if they
have been sent during the player's next turn. This is unacceptable, so
we must handle them gracefully and give adequate replies (for now, these
may be just error messages, as there is currently no meaningful action a
client might do outside of their turn). Since the overall architecture
of asyncronous control flow has been designed with this necessity in
mind, it should not be difficult to implement this feature.