Add a hijacking point in Server._handle_client().
After a stream has been created out of the socket in Server._handle_client,
it is passed to Server._handle_client_with_stream, which does the actual work
previously undertaken by the former method. However, it does not require
`stream` to be of the type `curio.io.SocketStream`, as it was the case
(de-facto) before the change. This will allow us to "hijack" this method by
replacing a real TCP stream with a mock version in tests.